[schema-registry-avro] Clean up tests and dependencies
