+++ sdk/schemaregistry/schema-registry-avro/rollup.test.config.js
@@ -1,3 +1,11 @@
-import * as base from "./rollup.base.config";
+import { makeConfig } from "@azure/dev-tool/shared-config/rollup";
+import * as base from "./rollup.patched.config";
 
-export default [base.nodeConfig(true), base.browserConfig(true)];
+const inputs = makeConfig(require("./package.json"));
+
+if (!process.env.ONLY_NODE) {
+  // replace the original test config with a patched one
+  inputs[1] = base.makeBrowserTestConfigPatchProcess();
+}
+
+export default inputs;
